// origin/src/youtube_music/types/ExploreResults_0.ts
export type ExploreResults_0 = Root2[]

export interface Root2 {
    responseContext: ResponseContext
    items?: Item[]
    trackingParams: string
    contents: Contents
    maxAgeStoreSeconds?: number
}

export interface ResponseContext {
    serviceTrackingParams: ServiceTrackingParam[]
}

export interface ServiceTrackingParam {
    service: string
    params: Param[]
}

export interface Param {
    key: string
    value: string
}

export interface Item {
    guideSectionRenderer: GuideSectionRenderer
}
// ...
export interface Contents {
    singleColumnBrowseResultsRenderer: SingleColumnBrowseResultsRenderer
}